\name{print.survNRI}

\alias{print.survNRI}

\title{

print an object created by the function survNRI
}
\description{
  S3 method for class ''survNRI''
}
\usage{
#S3 method for class ''survNRI''
\method{print}{survNRI}(x, ...)
}

\arguments{
  \item{x}{
   object of class "survNRI", output from the survNRI function 
}
\item{\dots}{ ignored}
}
